[refactor][l]: refactored code separating helpers functions
